microservices/flask_docs/requirements.txt
flask
flask_sqlalchemy
python-dotenv
itsdangerous
fpdf2
pymysql
